notes_backend/requirements.txt
asgiref==3.8.1
Django==4.2
djangorestframework==3.15.1
drf-yasg==1.21.7
inflection==0.5.1
packaging==24.0
pillow==10.3.0
psycopg2-binary==2.9.9
pytz==2024.1
PyYAML==6.0.1
setuptools==70.0.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
